Reorganized project structure and files
In this commit, I have reorganized the project files and structure for better organization and clarity. I removed the 'exercises.iml' file and introduced a new 'java-certified.iml' file. Also added '.idea' to '.gitignore' to exclude IDE specific files. Furthermore, commonly used classes User.java, Vars.java, and Main.java have been moved into a new package called 'exercises'. On top of this, I have updated the references to these classes in the code accordingly. This change will make it easier to expand the project with new modules and packages in the future.