feat: Add inheritance exercises and update README
Added several new classes 'Chair', 'Dresser', 'Lamp', 'Main', 'Product', 'Sofa', and 'Table' under 'src/excercises/inheritance' directory for inheritance exercises. These classes demonstrate the concept of inheritance in Java where 'Product' is a superclass and the others are subclasses. 'Main' class is used to test the functionalities.

Additionally, an error in the README linking to the relevant exercise has been corrected, and a link to the new inheritance exercises has been added.